# Conflicts:
#	wecom-objects/src/main/java/cn/felord/convert/EventEnumConverter.java
#	wecom-objects/src/main/java/cn/felord/enumeration/CallbackEvent.java

+++ wecom-objects/src/main/java/cn/felord/domain/agent/AllowParties.java
@@ -24,5 +24,5 @@ import java.util.List;
  */
 @Data
 public class AllowParties {
-    private List<Integer> partyid;
+    private List<Long> partyid;
 }
